# bridge/openapi_snapshot.json
{
  "components": {
    "schemas": {
      "Constraints": {
        "additionalProperties": false,
        "properties": {
          "allow_online_enrichment": {
            "default": false,
            "title": "Allow Online Enrichment",
            "type": "boolean"
          },
          "keep_language": {
            "default": true,
            "title": "Keep Language",
            "type": "boolean"
          },
          "max_chars": {
            "anyOf": [
              {
                "maximum": 200000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Chars"
          },
          "preserve_code_blocks": {
            "default": true,
            "title": "Preserve Code Blocks",
            "type": "boolean"
          }
        },
        "title": "Constraints",
        "type": "object"
      },
...
      "DiffInfo": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "title": "Data"
          },
          "type": {
            "default": "unified",
            "enum": ["spans", "unified"],
            "title": "Type",
            "type": "string"
          }
        },
        "required": ["data"],
        "title": "DiffInfo",
        "type": "object"
      },
      "ErrorItem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": ["code", "message"],
        "title": "ErrorItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Lane": {
        "enum": ["OB_OFFLINE", "OB_ONLINE_PULSE"],
        "title": "Lane",
        "type": "string"
      },
      "Mode": {
        "enum": ["clarify", "structure", "compress", "persona", "safe_exec"],
        "title": "Mode",
        "type": "string"
      },
      "MutationInfo": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "title": "Note",
            "type": "string"
          },
          "type": {
            "enum": [
              "reorder",
              "clarify",
              "add_constraints",
              "compress",
              "persona",
              "safety",
              "profile"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": ["type", "note"],
        "title": "MutationInfo",
        "type": "object"
      },
...
      "PolicyInfo": {
        "additionalProperties": false,
        "properties": {
          "lane_allowed": {
            "title": "Lane Allowed",
            "type": "boolean"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "network_used": {
            "title": "Network Used",
            "type": "boolean"
          }
        },
        "required": ["lane_allowed", "network_used", "model"],
        "title": "PolicyInfo",
        "type": "object"
      },
      "PromptContext": {
        "additionalProperties": false,
        "properties": {
          "intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "tab_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tab Title"
          },
          "tab_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tab Url"
          },
          "target_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Base Url"
          }
        },
        "title": "PromptContext",
        "type": "object"
      },
      "PromptEnhanceRequest": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "v1",
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "constraints": {
            "$ref": "#/components/schemas/Constraints"
          },
          "context": {
            "$ref": "#/components/schemas/PromptContext"
          },
          "lane": {
            "$ref": "#/components/schemas/Lane"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "prompt": {
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "schema_version": {
            "default": "2026-02-06",
            "title": "Schema Version",
            "type": "string"
          },
          "trace": {
            "$ref": "#/components/schemas/Trace"
          }
        },
        "required": ["lane", "prompt", "mode"],
        "title": "PromptEnhanceRequest",
        "type": "object"
      },
      "PromptEnhanceResponse": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "v1",
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "diff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiffInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "enhanced_prompt": {
            "title": "Enhanced Prompt",
            "type": "string"
          },
          "enhancement_id": {
            "title": "Enhancement Id",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "lane": {
            "title": "Lane",
            "type": "string"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "mutations": {
            "items": {
              "$ref": "#/components/schemas/MutationInfo"
            },
            "title": "Mutations",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "original_prompt": {
            "title": "Original Prompt",
            "type": "string"
          },
          "policy": {
            "$ref": "#/components/schemas/PolicyInfo"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "rationale": {
            "items": {
              "type": "string"
            },
            "title": "Rationale",
            "type": "array"
          },
          "redactions": {
            "items": {
              "$ref": "#/components/schemas/RedactionInfo"
            },
            "title": "Redactions",
            "type": "array"
          },
          "safety_flags": {
            "items": {
              "type": "string"
            },
            "title": "Safety Flags",
            "type": "array"
          },
          "schema_version": {
            "default": "2026-02-06",
            "title": "Schema Version",
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/Stats"
          }
        },
        "required": [
          "ok",
          "enhancement_id",
          "original_prompt",
          "enhanced_prompt",
          "mode",
          "lane",
          "policy",
          "stats",
          "provenance"
        ],
        "title": "PromptEnhanceResponse",
        "type": "object"
      },
      "Provenance": {
        "additionalProperties": false,
        "properties": {
          "bridge_version": {
            "title": "Bridge Version",
            "type": "string"
          },
          "generated_utc": {
            "title": "Generated Utc",
            "type": "string"
          },
          "retrieval_mode": {
            "title": "Retrieval Mode",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "generated_utc",
          "bridge_version",
          "retrieval_mode",
          "run_id"
        ],
        "title": "Provenance",
        "type": "object"
      },
      "RedactionInfo": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "title": "Note",
            "type": "string"
          },
          "type": {
            "enum": ["secret_like", "credential", "token"],
            "title": "Type",
            "type": "string"
          }
        },
        "required": ["type", "note"],
        "title": "RedactionInfo",
        "type": "object"
      },
      "Stats": {
        "additionalProperties": false,
        "properties": {
          "chars_after": {
            "title": "Chars After",
            "type": "integer"
          },
          "chars_before": {
            "title": "Chars Before",
            "type": "integer"
          },
          "est_tokens_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Tokens After"
          },
          "est_tokens_before": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Tokens Before"
          }
        },
        "required": ["chars_before", "chars_after"],
        "title": "Stats",
        "type": "object"
      },
      "Trace": {
        "additionalProperties": false,
        "properties": {
          "client_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Version"
          },
          "source": {
            "default": "openbrowser",
            "enum": ["openbrowser", "mcp", "cli"],
            "title": "Source",
            "type": "string"
          }
        },
        "title": "Trace",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SOCA OpenBrowser Bridge",
    "version": "v1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/soca/promptbuddy/enhance": {
      "post": {
        "operationId": "promptbuddy_enhance_soca_promptbuddy_enhance_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptEnhanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptEnhanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promptbuddy Enhance"
      }
    },
...
    "/soca/promptbuddy/profiles": {
      "get": {
        "operationId": "promptbuddy_profiles_soca_promptbuddy_profiles_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Promptbuddy Profiles Soca Promptbuddy Profiles Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promptbuddy Profiles"
      }
    },
    "/soca/promptbuddy/selftest": {
      "get": {
        "operationId": "promptbuddy_selftest_soca_promptbuddy_selftest_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Promptbuddy Selftest Soca Promptbuddy Selftest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promptbuddy Selftest"
      }
    },
...
  }
}